requirements/dev.txt
-r common.txt

bumpversion==0.5.3
django-debug-toolbar==1.9.1
django-nose==1.4.5
Fabric3==1.14.post1
factory-boy==2.10.0
nose==1.3.7
pylama==7.4.3
